requirements.txt
python_telegram_bot==12.5.1
pyautogui
pillow
